% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_content_from_hl7.R
\name{extract_content_from_hl7}
\alias{extract_content_from_hl7}
\title{extract content from a hl7}
\usage{
extract_content_from_hl7(file)
}
\arguments{
\item{file}{the path of the hl7 file}
}
\value{
a list with headers and their corresponding elements
}
\description{
This function will extract content of the hl7. However, the content will not have the lable of the elements
}
