% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_content_from_hl7_files.R
\name{extract_content_from_hl7_files}
\alias{extract_content_from_hl7_files}
\title{extract content from multiple hl7 files}
\usage{
extract_content_from_hl7_files(files, location)
}
\arguments{
\item{files}{list of files}

\item{location}{the path of the hl7 files, in case of}
}
\value{
a data.frame with headers and their corresponding elements with a sequence of column names, or NULL in case of no hl7 files
}
\description{
This function will extract content of the hl7. However, the content will not have the lable of the elements
}
