% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_HL7.R
\name{extract_HL7}
\alias{extract_HL7}
\title{Extracts information from hl7 file}
\usage{
extract_HL7(file)
}
\arguments{
\item{file}{is the path of the file}
}
\value{
returns a data.frame with column names MANDT, PATID, PATNR, FALNR, LFDNR, BEWTY, BWART, TITLE,
VORNAME,NACHNAME, GEBDAT, GSCHL, FAMST, KLASSE, ORGPF, ORGFA, DATUM, UHRZEIT.
}
\description{
extracts the data from hl7 file and return a data.frame with a set of defined column names.
}
