% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_time.R
\name{get_date_time}
\alias{get_date_time}
\title{reads date and time}
\usage{
get_date_time(str)
}
\arguments{
\item{str}{expects an date and time as a string in the format %Y%m%d%H%M%S
datetime <- '2010-02-13 10:30:00'
get_date_time(datetime)}
}
\description{
splits the string into date and time.
}
