% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hl7_data.R
\name{get_hl7_data}
\alias{get_hl7_data}
\title{extracts the content of the hl7 file}
\usage{
get_hl7_data(file, data)
}
\arguments{
\item{file}{path of the hl7 file}

\item{data}{content of the hl7 file}
}
\description{
reads either file or just the content of the file and returns a list of headers. The input shall be either file or data.
}
