% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hl7_files.R
\name{get_hl7_files}
\alias{get_hl7_files}
\title{gets all hl7 files from a path}
\usage{
get_hl7_files(files_location)
}
\arguments{
\item{files_location}{the path where the hl7 files exists}
}
\description{
lists all the hl7 files in a location
}
