% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hl7_line.R
\name{get_hl7_line}
\alias{get_hl7_line}
\title{get the details of the hl7 content based on the header}
\usage{
get_hl7_line(data, term)
}
\arguments{
\item{data}{hl7 content}

\item{term}{the header}
}
\description{
get the details of the hl7 content based on the header
}
