% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PID.R
\name{get_PID}
\alias{get_PID}
\alias{Sai}
\alias{Pavan}
\alias{Kumar}
\alias{Veeranki}
\title{Reads PID line from hl7}
\usage{
get_PID(data)
}
\arguments{
\item{data}{text from hl7 file}
}
\value{
a list with the elements from PID header
}
\description{
This function gets the data from PID line of HL7 file
}
