% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hl7_file.R
\name{read_hl7_file}
\alias{read_hl7_file}
\title{reads date hl7 file}
\usage{
read_hl7_file(file)
}
\arguments{
\item{file}{path of the hl7 file

reads whole hl7 file and returns the content of it ####}
}
\description{
reads date hl7 file
}
