% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_evn.R
\name{write_evn}
\alias{write_evn}
\title{Prepares EVENT with random information as a list}
\usage{
write_evn()
}
\value{
returns a list of list
}
\description{
generates a list of lists, with evn that cotains the information
from EVN and the line as its should be written in the .hl7 file.
}
