% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hl7s.R
\name{write_hl7s}
\alias{write_hl7s}
\title{writes hl7s with random data.}
\usage{
write_hl7s(write_to_file = TRUE, nr_files, return = TRUE)
}
\arguments{
\item{write_to_file}{is a binary input TRUE/FALSE with TRUE as default, FALSE prints to screen}

\item{nr_files}{is an integer greater than 0, if missing, function writes single file.}

\item{return}{is a binary input TRUE/FALSE. TRUE would deliver an input a list}
}
\value{
returns a data.frame with hl7 data with a next line  as a seperator
}
\description{
write_hl7s function writes hl7 files with random data that generates
for A09 transactions. This function writes the headers MSH, EVENT, PID, PV1, PV2
}
