% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msh.R
\name{write_msh}
\alias{write_msh}
\title{Prepares MSH with random information as a list}
\usage{
write_msh()
}
\value{
returns a list of list
}
\description{
generates a list of lists, with msh that cotains the information
from MSH and the line as its should be written in the .hl7 file.
}
