% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pid.R
\name{write_pid}
\alias{write_pid}
\title{Prepares PID with random information as a list}
\usage{
write_pid()
}
\value{
returns a list of list
}
\description{
generates a list of lists, with pid that cotains the information
from PID and the line as its should be written in the .hl7 file.
}
