% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pv1.R
\name{write_pv1}
\alias{write_pv1}
\title{Prepares PV1 with random information as a list}
\usage{
write_pv1()
}
\value{
returns a list of list
}
\description{
generates a list of lists, with pv1 that cotains the information
from PV1 and the line as its should be written in the .hl7 file.
}
