% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pv2.R
\name{write_pv2}
\alias{write_pv2}
\title{Prepares PV2 with random information as a list}
\usage{
write_pv2()
}
\value{
returns a list of list
}
\description{
generates a list of lists, with pv2 that cotains the information
from PV2 and the line as its should be written in the .hl7 file.
}
